Workaround list view selection oddity.
This problem would present itself when clicking on an item that
was partially in view at the bottom of the list. After the click
the wrong item was selected. Qt seems to delay their selection
update code by 1/2 second or something and this gets interfered
with when we set the selection immediately.

svn r964