Restructured Operation to separate data from processing.
Operation properties are now stored in a values dict
instead of directly in the operation. This means inspectors
and other manipulations must occur through standardized
functions.

Updated the database version to 2. This migration throws
out operation parameters -- but this shouldn't present
a problem for the few existing databases.

svn r1374