Add some permanent and temporary performance improvements.
Data calculation is only performed one at a time (serialized). Likewise
thumbnail and histograms are serialized too. And thumbnails and
histograms get delayed. Display loops are changed to periodic updates
rather than immediate.

Data range is problematic in that the inspector requires it to be
calculated in the UI thread if it doesn't exist. Temporarily changed it
so that it only calculates if it has no cached value. This results in
incorrect data range in many cases. Will be fixed in future version.

svn r1553