Deleting dependent items no longer requires db data_reference entry.
Add test to make sure deleting dependent item after source item was
deleted succeeds. The problem had been that the dependent item had
no entry in data_reference in the db, so it wouldn't get deleted.
The new way adds an item in the data_reference table, even for dependent
items; but no longer requires that entry to do the deletion. This makes
the db more independent of the data item files.

svn r1892