Untangle and separate document controller from library item list.
In the data panel, the item controllers (just controlling titles at
the moment) for the list of top level libraries is now separated
from the model controller. This is part of the change to minimize
the interactions between the various data panel pieces and the
document controller and model.

svn r2727