Restructure DisplayPanel so that content can be changed (live, browser, etc.)
This is similar to previous functionality, except that another layer is
introduced to make it possible to change the larger functionality of a
display panel. Previously only an 'image' display panel was supported
and the display controller for that display panel could be changed. Now
multiple display panel types are supported. Changing the display
controller also changes the display panel content object instead of just
re-using the previous one.

This is another step towards generalizing display panels to allow for
browsers and other content.

svn r3174