Begin separating data panel functionality into UI and controller.
Created a DataBrowserController, but in doing so it is clear that
update_data_panel_selection is a problematic method. This
commit should be considered partial code on the road to
getting rid of update_data_panel_selection, which will happen
by moving the selection handling to the document controller.

svn r3185