Fix problems with shared selection object.
Multiple entities were modifying the selection in the case of
insert/remove data items. This led to inconsistent selection
handling.

In addition, there are structural problems when deleting multiple
data items. The problem arises when the display items aren't
updated until periodic is called; but there is also code to use
display items when the selected item is deleted. This commit
includes a hack to get around that problem by calling periodic
during the delete loop.

svn r3438