Fix case of inspector out of sync with deleted data item.
This is a messy situation and the solution is equally messy.
But this handles the case where the inspector gets updated
from a thread and the data item gets deleted before the
inspector can be serviced on the main thread.

Also added stub tests for a few more corrupt data item issues.

svn r3455