Fix problem with inspector updating on background thread.
Causes crash when used with Swift.exe. The problem is
occurring because when the probe marker is selected, the
inspector for it is active. The inspector watches for changes
and updates if the probe marker disappears (which happens
when scanning starts). Unfortunately, the probe marker
disappears on a background thread, causing the inspector
to update on a background thread, which calls back to
Swift.exe and throws an exception which causes the crash.

Fixes #485

svn r3738