Improve computation queue to deal with data items being removed.
This avoids usage of None when a dependent data item is deleted
during live acquisition.

Also added a few tests which partially test for the problem,
although the tests can pass and the problem can still exist on
a live system due to threading behavior differences between
the tests and live application.