Change histogram/stats to use async streams.
While the code is still somewhat more complicated that desired,
it is simpler in the sense that it no longer uses threads but
instead relies on asyncio. Actual calculation of stats and histogram
is still technically done in a thread, but via asyncio and much
simpler to understand. Tests are now more reliable too.