Switch to using API data item in computations (partial).
This is part of a change to standardize the computations to allow
users to use the standard API where possible, use standard Python
where possible, and to allow computations more flexibility in how
they update data items.

This commit does not include migration of the file format and
older data items (v10) with computations will not load.