Added "Checkbox Script" section
I looked at the checkboxes.py script and wrote up this documentation. I was not quite sure exactly which part of the documentation you wanted this included, so I added it with the other scripting information. Just let me know if you have any questions or concerns about it. Thank you!

-Victor