Fix another selection/focus issue.
This one had to do with the displayed item in a panel disappearing
when a filter was applied to the data panel, causing the displayed
item to no longer be available in the display panel browser.

The fix was to distinguish between changes made by the user directly
to the display panel and those changes occuring implicitly by other
means.