Made adding a stack to a graph only display an outline of the added graph
Modified the uniqiue color fethching algorithim to check for colors without transparency.
Modified the layer adding functionality to check if an added layer is part of a stack.
If a layer is part of a stack, it will now add transparency to the fill color.
Additionally modified the layer adding functionality to give each graph
an outline of the same fill color it used by default.