Fix issue with how listeners get closed (use Python ref counting instead of explicit close call).
Closing event listeners is a null operation. Instead all Python references must be removed.
This change ensures that the listener itself doesn't keep a reference to the model, which would
prevent it from being garbage collected.

This could be reproduced by making a line profile, adding an interval region to the resulting plot,
moving it, then moving the line profile graphic on the raster image and observing log messages.

Added a test, but it's a difficult error to check for so the test is more like a confirmation.
Numerous incorrect changes to the code will still allow this test to pass.