Adjusting Test to check accuracy
Added a new function to verify Tuples are 'close', with a provided delta or precision.
Changed the unit test to work, as well as checking the mouse position is approximately correct.  Future Issue needs to verify rendering precision on the layout which may be introducing inaccuracies with mouse->image offsets.