Refactor ZoomMouseHandler
Refactored the zoom function to not be a mousehandler.
Consolidated functions to remove duplicate code, so now zoom_in and zoom_out call down to a single zoom routine.
Fixed an error in the calculation of what zoom level the various display modes are in when switching to custom.
Consolidated the display_properties_command into a single call at the end once all parameters are known.