Fix #1203. Eliminate use of data sampling to calculate complex display limits.
The sample size was 200 and in no way could be representative of an
arbitrary complex valued image displayed in log-absolute. This was probably
done for perceived performance; but modern processors make the performance
issue minimal. So do not sample; just use the full range until a good
general sampling solution is implemented.