Add Verify of Project Name on text_changed
Added explicit sanity check of Project Name when text changed on the field.  Verifies compatibility using Utility.simplify_filename which detects any non-compatible filenames and returns a sanitised version so if they are different it means it was not compatible.
Currently the identification for the user is some orange background flair and a tooltip on the Create button.