Moved Edit button to middle, and changed icon to a chevron
This is to allow for UX preview testing, with icon and location changing.

This has one outstanding bug, and one feature deviation.  The bug is handling of repeated pressing of the edit button causes the popup to reload.  The deviation is the tooltip/drag is now on the title rather than the entire bar.