Generalize processing computation to return multiple outputs, specify input data types.
The processing computation is a computation handler which takes a
processing base subclass and applies it to a collection point by
point. This changes it to also register outputs and stricter inputs.

As a side effect, it also fixes an issue would trigger processing
computations to recompute when changing the display, resulting in
sometimes very slow unnecessary recomputes. A test for this is
included in an upcoming commit.