Simplify processing base objects to take data and metadata directly.
Eliminates ProcessingDataSource which was facilitating late
computation of filtered data. The new method declares that the
filtered data should be passed to the processing base directly. This
ensures that changing the filter will efficiently trigger recompute.

Add a new method to get the input data for a given index for use in
processing base.